MODOAIPMH-509 - [Orchid] [OAI-PMH] Full harvesting - OutOfMemory Exception
